requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
altgraph==0.17.3
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
auto-py-to-exe==2.23.1
backcall==0.2.0
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
debugpy==1.6.3
decorator==5.1.1
discord.py==2.1.0
Django==4.1.3
Eel==0.14.0
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
frozenlist==1.3.3
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
hikari==2.0.0.dev112
hikari-lightbulb==2.2.5
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.4
jupyter_core==5.0.0
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pefile==2022.5.30
pi==0.1.2
pickleshare==0.7.5
platformdirs==2.5.3
prompt-toolkit==3.0.32
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.6.0
tornado==6.2
traitlets==5.5.0
tzdata==2022.7
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
whichcraft==0.6.1
WTForm==1.0
WTForms==3.0.1
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.1
